Build Object

Build Object

Build(
  Build.substitutions = NULL,
  Build.timing = NULL,
  results = NULL,
  logsBucket = NULL,
  steps = NULL,
  buildTriggerId = NULL,
  id = NULL,
  tags = NULL,
  startTime = NULL,
  substitutions = NULL,
  timing = NULL,
  sourceProvenance = NULL,
  createTime = NULL,
  images = NULL,
  projectId = NULL,
  logUrl = NULL,
  finishTime = NULL,
  source = NULL,
  options = NULL,
  timeout = NULL,
  status = NULL,
  statusDetail = NULL,
  artifacts = NULL,
  secrets = NULL
)

Arguments

Build.substitutions
The Build.substitutions object or list of objects

Build.timing
The Build.timing object or list of objects

results
Output only

logsBucket
Google Cloud Storage bucket where logs should be written (see

steps
Required

buildTriggerId
Output only

id
Output only

tags
Tags for annotation of a `Build`

startTime
Output only

substitutions
Substitutions data for `Build` resource

timing
Output only

sourceProvenance
Output only

createTime
Output only

images
A list of images to be pushed upon the successful completion of all build

projectId
Output only

logUrl
Output only

finishTime
Output only

source
A Source object specifying the location of the source files to build, usually created by cr_build_source

options
Special options for this build

timeout
Amount of time that this build should be allowed to run, to second

status
Output only

statusDetail
Output only

artifacts
Artifacts produced by the build that should be uploaded upon

secrets
Secrets to decrypt using Cloud Key Management Service

Value

Build object

Details

A build resource in the Cloud Build API.

At a high level, a `Build` describes where to find source code, how to build it (for example, the builder image to run on the source), and where to store the built artifacts.

Build Macros

Fields can include the following variables, which will be expanded when the build is created:-

$PROJECT_ID: the project ID of the build.

$BUILD_ID: the autogenerated ID of the build.

$REPO_NAME: the source repository name specified by RepoSource.

$BRANCH_NAME: the branch name specified by RepoSource.

$TAG_NAME: the tag name specified by RepoSource.

$REVISION_ID or $COMMIT_SHA: the commit SHA specified by RepoSource or resolved from the specified branch or tag.

$SHORT_SHA: first 7 characters of $REVISION_ID or $COMMIT_SHA.

See also