Job Schedule Object

Job Schedule Object

Job(attemptDeadline = NULL, pubsubTarget = NULL, httpTarget = NULL,
  timeZone = NULL, description = NULL, appEngineHttpTarget = NULL,
  status = NULL, retryConfig = NULL, state = NULL, name = NULL,
  lastAttemptTime = NULL, scheduleTime = NULL, schedule = NULL,
  userUpdateTime = NULL)

Arguments

attemptDeadline
The deadline for job attempts

pubsubTarget
Pub/Sub target

httpTarget
A HTTP target object HttpTarget

timeZone
Specifies the time zone to be used in interpreting

description
Optionally caller-specified in CreateJob or

appEngineHttpTarget
App Engine HTTP target

status
Output only

retryConfig
Settings that determine the retry behavior

state
Output only

name
Optionally caller-specified in CreateJob, after

lastAttemptTime
Output only

scheduleTime
Output only

schedule
A cron schedule e.g. "15 5 * * *"

userUpdateTime
Output only

Value

Job object

Details

Configuration for a job.The maximum allowed size for a job is 100KB.

See also