Set up Google Cloud Build to run a targets pipeline

Creates a Google Cloud Build yaml file so as to execute tar_make pipelines

Historical runs accumulate in the configured Google Cloud Storage bucket, and the latest output is downloaded before tar_make executes so up-to-date steps do not rerun.

cr_build_targets(
  task_image = "gcr.io/gcer-public/targets",
  target_folder = NULL,
  path = "cloudbuild_targets.yaml",
  bucket = cr_bucket_get(),
  task_args = list(),
  tar_make = "targets::tar_make()",
  ...
)

cr_build_targets_artifacts(
  build,
  download_folder = "_targets_cloudbuild",
  target_subfolder = c("all", "meta", "objects", "user"),
  overwrite = TRUE
)

Arguments

task_image

An existing Docker image that will be used to run your targets workflow after the targets meta has been downloaded from Google Cloud Storage

target_folder

Where target metadata will sit within the Google Cloud Storage bucket as a folder. If NULL defaults to RStudio project name or "targets_cloudbuild" if no RStudio project found.

path

File path to write the Google Cloud Build yaml workflow file. Set to NULL to write no file and just return the Yaml object.

bucket

The Google Cloud Storage bucket the target metadata will be saved to in folder `target_folder`

task_args

A named list of additional arguments to send to cr_buildstep_r when its executing the tar_make command (such as environment arguments)

tar_make

The R script that will run in the tar_make() step. Modify to include custom settings

...

Arguments passed on to cr_build_yaml

steps

A vector of cr_buildstep

timeout

How long the entire build will run. If not set will be 10mins

logsBucket

Where logs are written. If you don't set this field, Cloud Build will use a default bucket to store your build logs.

options

A named list of options

substitutions

Build macros that will replace entries in other elements

tags

Tags for the build

secrets

A secrets object

images

What images will be build from this cloudbuild

artifacts

What artifacts may be built from this cloudbuild - create via cr_build_yaml_artifact

availableSecrets

What environment arguments from Secret Manager are available to the build - create via cr_build_yaml_secrets

serviceAccount

What service account should the build be run under?

build

A Build object that includes the artifact location

download_folder

Where to download the artifact files

target_subfolder

If you only want to download a specific folder from the _targets/ folder on Cloud Build then specify it here.

overwrite

Whether to overwrite existing local data

Value

A Yaml object as generated by cr_build_yamlcr_build_targets_artifacts returns the file path to where the download occurred.

Details

Steps to set up your target task in Cloud Build:

Create your `targets` workflow.

Create a Dockerfile that holds the R and system dependencies for your workflow. You can test the image using cr_deploy_docker. Include library(targets) dependencies - a Docker image with targets installed is available at gcr.io/gcer-public/targets.

Run cr_build_targets to create the cloudbuild yaml file.

Run the build via cr_build or similar. Each build should only recompute outdated targets.

Optionally create a build trigger via cr_buildtrigger.

Trigger a build. The first trigger will run the targets pipeline, subsequent runs will only recompute the outdated targets.

Use cr_build_targets_artifacts to download the return values of a target Cloud Build, then tar_read to read the results. You can set the downloaded files as the target store via targets::tar_config_set(store="_targets_cloudbuild"). Set download_folder = "_targets" to overwrite your local targets store.

See also

Examples

cr_build_targets()
#> ℹ 2021-12-20 08:29:11 > targets cloud location: gs://my-bucket/targets_cloudbuild
#> ℹ 2021-12-20 08:29:11 > Writing to cloudbuild_targets.yaml
#> ==cloudRunnerYaml==
#> steps:
#> - name: gcr.io/google.com/cloudsdktool/cloud-sdk:alpine
#>   entrypoint: bash
#>   args:
#>   - -c
#>   - mkdir /workspace/_targets && mkdir /workspace/_targets/meta && gsutil -m cp -r
#>     ${_TARGET_BUCKET}/_targets/meta /workspace/_targets || exit 0
#>   id: get previous _targets metadata
#> - name: gcr.io/gcer-public/targets
#>   args:
#>   - Rscript
#>   - -e
#>   - targets::tar_make()
#>   id: target pipeline
#> - name: gcr.io/google.com/cloudsdktool/cloud-sdk:alpine
#>   entrypoint: bash
#>   args:
#>   - -c
#>   - date > buildtime.txt && gsutil cp buildtime.txt ${_TARGET_BUCKET}/_targets/buildtime.txt
#>   id: Ensure bucket/_targets/ always exists
#> - name: gcr.io/google.com/cloudsdktool/cloud-sdk:alpine
#>   entrypoint: bash
#>   args:
#>   - -c
#>   - gsutil -m cp -r /workspace/_targets ${_TARGET_BUCKET} && gsutil ls -r ${_TARGET_BUCKET}
#>   id: Upload Artifacts
#> timeout: 3600s
#> substitutions:
#>   _TARGET_BUCKET: gs://my-bucket/targets_cloudbuild
unlink("cloudbuild_targets.yaml")