Create a StorageSource

This creates a StorageSource object after uploading to Google Cloud Storage

cr_build_upload_gcs(
  local,
  remote = paste0(local, format(Sys.time(), "%Y%m%d%H%M%S"), ".tar.gz"),
  bucket = cr_bucket_get(),
  predefinedAcl = "bucketOwnerFullControl",
  deploy_folder = "deploy"
)

Arguments

local
Local directory containing the Dockerfile etc. you wish to deploy

remote
The name of the folder in your bucket

bucket
The Google Cloud Storage bucket to upload to

predefinedAcl
The ACL rules for the object uploaded. Set to "bucketLevel" for buckets with bucket level access enabled

deploy_folder
Which folder to deploy from

Value

A Source object

Details

It copies the files into a folder call "deploy" in your working directory, then tars it for upload

See also

Examples

if (FALSE) {
cr_project_set("my-project")
cr_bucket_set("my-bucket")
my_gcs_source <- cr_build_upload_gcs("my_folder")
build1 <- cr_build("cloudbuild.yaml", source = my_gcs_source)

}