Create a build step to build and push a docker image

Create a build step to build and push a docker image

cr_buildstep_docker(
  image,
  tag = "$BUILD_ID",
  location = ".",
  projectId = cr_project_get(),
  ...
)

Arguments

image
The image tag that will be pushed, starting with gcr.io or created by combining with projectId if not starting with gcr.io

tag
The tag to attached to the pushed image - can use Build macros

location
Where the Dockerfile to build is in relation to dir

projectId
The projectId

...
Further arguments passed in to cr_buildstep

See also

Examples

cr_project_set("my-project")

#> 2020-02-14 11:24:34> ProjectId set to my-project

#> [1] "my-project"

cr_bucket_set("my-bucket")

#> 2020-02-14 11:24:34> Bucket set to my-bucket

#> [1] "my-bucket"

cr_buildstep_docker("gcr.io/my-project/my-image")

#> 2020-02-14 11:24:34> Image to be built: gcr.io/my-project/my-image:$BUILD_ID

#> [[1]]
#> ==cloudRunnerBuildStep==
#> name: gcr.io/cloud-builders/docker
#> args:
#> - build
#> - -t
#> - gcr.io/my-project/my-image:$BUILD_ID
#> - '.'
#> 
#> [[2]]
#> ==cloudRunnerBuildStep==
#> name: gcr.io/cloud-builders/docker
#> args:
#> - push
#> - gcr.io/my-project/my-image:$BUILD_ID
#> 

cr_buildstep_docker("my-image")

#> 2020-02-14 11:24:34> Image to be built: gcr.io/my-project/my-image:$BUILD_ID

#> [[1]]
#> ==cloudRunnerBuildStep==
#> name: gcr.io/cloud-builders/docker
#> args:
#> - build
#> - -t
#> - gcr.io/my-project/my-image:$BUILD_ID
#> - '.'
#> 
#> [[2]]
#> ==cloudRunnerBuildStep==
#> name: gcr.io/cloud-builders/docker
#> args:
#> - push
#> - gcr.io/my-project/my-image:$BUILD_ID
#> 

cr_buildstep_docker("my-image", tag = "$BRANCH_NAME")

#> 2020-02-14 11:24:34> Image to be built: gcr.io/my-project/my-image:$BRANCH_NAME

#> [[1]]
#> ==cloudRunnerBuildStep==
#> name: gcr.io/cloud-builders/docker
#> args:
#> - build
#> - -t
#> - gcr.io/my-project/my-image:$BRANCH_NAME
#> - '.'
#> 
#> [[2]]
#> ==cloudRunnerBuildStep==
#> name: gcr.io/cloud-builders/docker
#> args:
#> - push
#> - gcr.io/my-project/my-image:$BRANCH_NAME
#> 

# setting up a build to trigger off a Git source:
my_image <- "gcr.io/my-project/my-image"
my_repo <- RepoSource("github_markedmondson1234_googlecloudrunner",
                      branchName="master")
if (FALSE) {
docker_yaml <- cr_build_yaml(steps = cr_buildstep_docker(my_image))
built_docker <- cr_build(docker_yaml, source = my_repo)

# make a build trigger so it builds on each push to master
cr_buildtrigger("build-docker", trigger = my_repo, build = built_docker)
}