Create a build step for authenticating with Git

This creates steps to configure git to use an ssh created key.

This creates steps to use git with an ssh created key.

cr_buildstep_gitsetup(
  keyring = "my-keyring",
  key = "github-key",
  cipher = "id_rsa.enc",
  ...
)

cr_buildstep_git(
  git_args = c("clone", "git@github.com:[GIT-USERNAME]/[REPOSITORY]", "."),
  ...
)

Arguments

keyring
The Key Management Store keyring containing the git ssh key

key
The Key Management Store key containing the gitssh key

cipher
The filename of the encrypted git ssh key that has been checked into the repository

...
Further arguments passed in to cr_buildstep

git_args
The arguments to send to git

Details

The key should be encrypted offline using gcloud kms or similar first. See cr_buildstep_decrypt for details.

By default the encrypted key should then be at the root of your Source object called "id_rsa.enc"

cr_buildstep must come after cr_buildstep_gitsetup

See also

Examples

cr_project_set("my-project")

#> 2020-02-28 14:47:14> ProjectId set to my-project

#> [1] "my-project"

cr_bucket_set("my-bucket")

#> 2020-02-28 14:47:14> Bucket set to my-bucket

#> [1] "my-bucket"

# assumes you have previously saved git ssh key via KMS called "git_key"
cr_build_yaml(
     steps = c(
          cr_buildstep_gitsetup("my_keyring", "git_key"),
          cr_buildstep_git(c("clone",
                             "git@github.com:github_name/repo_name"))
     )
 )

#> ==cloudRunnerYaml==
#> steps:
#> - name: gcr.io/cloud-builders/gcloud
#>   args:
#>   - kms
#>   - decrypt
#>   - --ciphertext-file
#>   - id_rsa.enc
#>   - --plaintext-file
#>   - /root/.ssh/id_rsa
#>   - --location
#>   - global
#>   - --keyring
#>   - my_keyring
#>   - --key
#>   - git_key
#>   volumes:
#>   - name: ssh
#>     path: /root/.ssh
#> - name: gcr.io/cloud-builders/git
#>   entrypoint:
#>   - bash
#>   args:
#>   - -c
#>   - |
#>     chmod 600 /root/.ssh/id_rsa
#>     cat <<EOF >known_hosts
#>     github.com ssh-rsa AAAAB3NzaC1yc2EAAAABIwAAAQEAq2A7hRGmdnm9tUDbO9IDSwBK6TbQa+PXYPCPy6rbTrTtw7PHkccKrpp0yVhp5HdEIcKr6pLlVDBfOLX9QUsyCOV0wzfjIJNlGEYsdlLJizHhbn2mUjvSAHQqZETYP81eFzLQNnPHt4EVVUh7VfDESU84KezmD5QlWpXLmvU31/yMf+Se8xhHTvKSCZIFImWwoG6mbUoWf9nzpIoaSjB+weqqUUmpaaasXVal72J+UX2B+2RPW3RcT0eOzQgqlJL3RKrTJvdsjE3JEAvGq3lGHSZXy28G3skua2SmVi/w4yCE6gbODqnTWlg7+wC604ydGXA8VJiS5ap43JXiUFFAaQ==
#>     EOF
#>     cat <<EOF >/root/.ssh/config
#>     Hostname github.com
#>     IdentityFile /root/.ssh/id_rsa
#>     EOF
#>     mv known_hosts /root/.ssh/known_hosts
#>   volumes:
#>   - name: ssh
#>     path: /root/.ssh
#> - name: gcr.io/cloud-builders/git
#>   args:
#>   - clone
#>   - git@github.com:github_name/repo_name
#>   volumes:
#>   - name: ssh
#>     path: /root/.ssh