Create buildsteps to deploy to Cloud Run

Create buildsteps to deploy to Cloud Run

cr_buildstep_run(
  name,
  image,
  allowUnauthenticated = TRUE,
  region = cr_region_get(),
  concurrency = 80,
  port = NULL,
  max_instances = "default",
  memory = "256Mi",
  cpu = 1,
  env_vars = NULL,
  ...
)

Arguments

name
Name for deployment on Cloud Run

image
The name of the image to create or use in deployment - gcr.io

allowUnauthenticated
TRUE if can be reached from public HTTP address.

region
The endpoint region for deployment

concurrency
How many connections each container instance can serve. Can be up to 80.

port
Container port to receive requests at. Also sets the $PORT environment variable. Must be a number between 1 and 65535, inclusive. To unset this field, pass the special value "default".

max_instances
the desired maximum nuimber of container instances. "default" is 1000, you can get more if you requested a quota instance. For Shiny instances on Cloud Run, this needs to be 1.

memory
The format for size is a fixed or floating point number followed by a unit: G, M, or K corresponding to gigabyte, megabyte, or kilobyte, respectively, or use the power-of-two equivalents: Gi, Mi, Ki corresponding to gibibyte, mebibyte or kibibyte respectively. The default is 256Mi

cpu
1 or 2 CPUs for your instance

env_vars
Environment arguments passed to the Cloud Run container at runtime. Distinct from env that run at build time.

...
passed on to cr_buildstep

See also