Buildstep to run a targets pipeline on Cloud Build

This is a buildstep to help upload a targets pipeline, see cr_build_targets for examples and suggested workflow

cr_buildstep_targets(
  task_args = NULL,
  tar_make = "targets::tar_make()",
  task_image = "gcr.io/gcer-public/targets",
  id = "target pipeline"
)

cr_buildstep_targets_setup(bucket_folder)

cr_buildstep_targets_teardown(bucket_folder, last_id = NULL)

Arguments

task_args

If not NULL, a named list of additional arguments to send to cr_buildstep_r when its executing the tar_make command (such as environment arguments or waitFor ids)

tar_make

The R script that will run in the tar_make() step. Modify to include custom settings

task_image

An existing Docker image that will be used to run your targets workflow after the targets meta has been downloaded from Google Cloud Storage

id

The id of the buildstep. In linkcr_buildstep_targets_multi this is used along with waitFor to determine the order of execution

bucket_folder

The Google Cloud Storage bucket and folder the target metadata will be saved to, e.g. gs://my-bucket/my_target_project You can also pass in build substitution variables such as "${_MY_BUCKET}".

last_id

The final buildstep that needs to complete before the upload. If left NULL then will default to the last tar_target step.

See also