Deploy a Dockerfile so it will be built on ContainerRegistry

If no Dockerfile present in the deployment folder, will attempt to create a Dockerfile to upload via cr_dockerfile

cr_deploy_docker(local, image_name = remote, dockerfile = NULL,
  remote = basename(local), tag = "$BUILD_ID", timeout = 600L,
  bucket = cr_bucket_get(), projectId = cr_project_get(),
  launch_browser = interactive(), task_id = NULL)

Arguments

local
The folder containing the Dockerfile to build

image_name
The name of the docker image to be built either full name starting with gcr.io or constructed from the image_name and projectId via gcr.io/{projectId}/{image_name}

dockerfile
An optional Dockerfile built to support the script. Not needed if 'Dockerfile' exists in folder. If supplied will be copied into deployment folder and called "Dockerfile"

remote
The folder on Google Cloud Storage

tag
The tag to attached to the pushed image - can use Build macros

timeout
Amount of time that this build should be allowed to run, to second

bucket
The GCS bucker that will be used to deploy code source

projectId
The projectId

launch_browser
Whether to launch the logs URL in a browser once deployed

task_id
RStudio job task_id if you want to use the same task

Examples

# NOT RUN {
cr_deploy_docker(system.file("example", package="googleCloudRunner"))

# }