Deploy Docker build from a GitHub repo (Experimental)

This helps the common use case of building a Dockerfile based on the contents of a GitHub repo, and sets up a build trigger so it will build on every commit.

cr_deploy_github_docker(
  x,
  image = x,
  branch = ".*",
  image_tag = "$SHORT_SHA",
  dockerfile_location = ".",
  github_tag = NULL,
  timeout = 600L,
  projectId = cr_project_get()
)

Arguments

x
The GitHub repo e.g. MarkEdmondson1234/googleCloudRunner

image
The name of the image you want to build

branch
A regex of the GitHub branches that will trigger a build

image_tag
What to tag the build docker image

dockerfile_location
Where the Dockerfile sits within the GitHub repo

github_tag
Regexes matching what tags to build. If not NULL then argument branch will be ignored

timeout
timeout for the Docker build

projectId
The project to build under

Details

Build trigger API is experimental so this function is in development.

See also