Deploy a cloudbuild.yml for R package tests and upload to Codecov

This tests an R package each time you commit, and uploads the test coverage results to Codecov

cr_deploy_packagetests(
  steps = NULL,
  cloudbuild_file = "cloudbuild-tests.yml",
  env = c("NOT_CRAN=true"),
  test_script = NULL,
  codecov_script = NULL,
  codecov_token = "$_CODECOV_TOKEN",
  build_image = "gcr.io/gcer-public/packagetools:master",
  ...
)

Arguments

steps
extra steps to run before the cr_buildstep_packagetests steps run (such as decryption of auth files)

cloudbuild_file
The cloudbuild yaml file to write to

env
Environment arguments to be set during the test script runs

test_script
The script that will run first making tests. If NULL a default script it used

codecov_script
The script that will run first making tests. If NULL a default script it used

codecov_token
If using codecov, supply your codecov token here. Default assumes you add it to the Cloud Build substitution macros, which is more secure and recommended.

build_image
The docker image that will be used to run the R code for the test scripts

...
Arguments passed on to cr_build_make

yaml

A Yaml object created from cr_build_yaml or a file location of a .yaml/.yml cloud build file

artifacts

Artifacts that may be built via cr_build_yaml_artifact

options

Options

source

A Source object specifying the location of the source files to build, usually created by cr_build_source

timeout

Amount of time that this build should be allowed to run, to second

images

A list of images to be pushed upon the successful completion of all build

substitutions

Substitutions data for `Build` resource

Details

The trigger repository needs to hold an R package configured to do tests upon.

For GitHub, the repository will need to be linked to the project you are building within, via https://console.cloud.google.com/cloud-build/triggers/connect

If your tests need authentication details, add these via cr_buildstep_decrypt to the steps argument, which will prepend decrypting the authentication file before running the tests.

If you want codecov to ignore some files then also deploy a .covrignore file to your repository - see covr website at https://covr.r-lib.org/ for details.

See also

Examples

pd <- cr_deploy_packagetests()

#> ℹ 2020-06-01 09:15:19 > Writing to  cloudbuild-tests.yml

#> 

#> ℹ Complete deployment of tests Cloud Build yaml:

#> ● Go to https://console.cloud.google.com/cloud-build/triggers and
#>   make a build trigger pointing at this file in your repo:
#>   cloudbuild-tests.yml 

#> ℹ Build Trigger substitution variable settings:
#>   _CODECOV_TOKEN = your-codecov-token
#>   Ignored files filter (glob): docs/** and vignettes/**

pd

#> ==cloudRunnerYaml==
#> steps:
#> - name: gcr.io/gcer-public/packagetools:master
#>   args:
#>   - Rscript
#>   - -e
#>   - |-
#>     message("cran mirror: ", getOption("repos"))
#>     remotes::install_deps(dependencies = TRUE)
#>     rcmdcheck::rcmdcheck(args = '--no-manual', error_on = 'warning')
#>   env:
#>   - NOT_CRAN=true
#> - name: gcr.io/gcer-public/packagetools:master
#>   args:
#>   - Rscript
#>   - -e
#>   - |-
#>     remotes::install_deps(dependencies = TRUE)
#>     remotes::install_local()
#>     cv <- covr::package_coverage()
#>     print(cv)
#>     covr::codecov(coverage=cv, commit = '$COMMIT_SHA', branch = '$BRANCH_NAME')
#>   env:
#>   - NOT_CRAN=true
#>   - CODECOV_TOKEN=$_CODECOV_TOKEN

# add a decryption step for an auth file
cr_deploy_packagetests(
  steps = cr_buildstep_secret("my_secret", "auth.json"),
  env = c("NOT_CRAN=true", "MY_AUTH_FILE=auth.json"),
  timeout = 1200
)

#> ℹ 2020-06-01 09:15:19 > Writing to  cloudbuild-tests.yml

#> 

#> ℹ Complete deployment of tests Cloud Build yaml:

#> ● Go to https://console.cloud.google.com/cloud-build/triggers and
#>   make a build trigger pointing at this file in your repo:
#>   cloudbuild-tests.yml 

#> ℹ Build Trigger substitution variable settings:
#>   _CODECOV_TOKEN = your-codecov-token
#>   Ignored files filter (glob): docs/** and vignettes/**

unlink("cloudbuild-tests.yml")