Deploy a cloudbuild.yml for a pkgdown website of an R package

This builds a pkgdown website each time the trigger fires and deploys it to git

cr_deploy_pkgdown(
  steps = NULL,
  cloudbuild_file = "cloudbuild-pkgdown.yml",
  git_email = "googlecloudrunner@r.com",
  keyring = "my-keyring",
  key = "github-key",
  env = NULL,
  cipher = "id_rsa.enc",
  build_image = "gcr.io/gcer-public/packagetools:master"
)

Arguments

steps
extra steps to run before the pkgdown website steps run

cloudbuild_file
The cloudbuild yaml file to write to

git_email
The email the git commands will be identifying as

keyring
The Key Management Store keyring containing the git ssh key

key
The Key Management Store key containing the gitssh key

env
A character vector of env arguments to set for all steps

cipher
The filename of the encrypted git ssh key that has been checked into the repository

build_image
A docker image with pkgdown installed

Details

The trigger repository needs to hold an R package configured to build a pkgdown website.

For GitHub, the repository will also need to be linked to the project you are building within, via https://console.cloud.google.com/cloud-build/triggers/connect

The git ssh keys need to be deployed to Google KMS for the deployment of the website - see cr_buildstep_git - this only needs to be done once per Git account. You then need to commit the encrypted ssh key (by default called id_rsa.enc)

See also

Examples

pd <- cr_deploy_pkgdown()

#> 2020-02-14 10:32:39> Writing to cloudbuild-pkgdown.yml

#> 2020-02-14 10:32:39> Now make a build trigger pointing at this file in your repo: cloudbuild-pkgdown.yml
#> Build Trigger settings:
#> Substitution variable: _GITHUB_REPO = {your-repo-to-push-to}
#> Ignored files filter (glob): docs/**

pd

#> ==CloudBuildObject==
#> steps:
#> - name: gcr.io/cloud-builders/gcloud
#>   args:
#>   - kms
#>   - decrypt
#>   - --ciphertext-file
#>   - id_rsa.enc
#>   - --plaintext-file
#>   - /root/.ssh/id_rsa
#>   - --location
#>   - global
#>   - --keyring
#>   - my-keyring
#>   - --key
#>   - github-key
#>   volumes:
#>   - name: ssh
#>     path: /root/.ssh
#> - name: gcr.io/cloud-builders/git
#>   entrypoint:
#>   - bash
#>   args:
#>   - -c
#>   - |
#>     chmod 600 /root/.ssh/id_rsa
#>     cat <<EOF >known_hosts
#>     github.com ssh-rsa AAAAB3NzaC1yc2EAAAABIwAAAQEAq2A7hRGmdnm9tUDbO9IDSwBK6TbQa+PXYPCPy6rbTrTtw7PHkccKrpp0yVhp5HdEIcKr6pLlVDBfOLX9QUsyCOV0wzfjIJNlGEYsdlLJizHhbn2mUjvSAHQqZETYP81eFzLQNnPHt4EVVUh7VfDESU84KezmD5QlWpXLmvU31/yMf+Se8xhHTvKSCZIFImWwoG6mbUoWf9nzpIoaSjB+weqqUUmpaaasXVal72J+UX2B+2RPW3RcT0eOzQgqlJL3RKrTJvdsjE3JEAvGq3lGHSZXy28G3skua2SmVi/w4yCE6gbODqnTWlg7+wC604ydGXA8VJiS5ap43JXiUFFAaQ==
#>     EOF
#>     cat <<EOF >/root/.ssh/config
#>     Hostname github.com
#>     IdentityFile /root/.ssh/id_rsa
#>     EOF
#>     mv known_hosts /root/.ssh/known_hosts
#>   volumes:
#>   - name: ssh
#>     path: /root/.ssh
#> - name: gcr.io/cloud-builders/git
#>   args:
#>   - clone
#>   - git@github.com:$_GIT_REPO
#>   - repo
#>   volumes:
#>   - name: ssh
#>     path: /root/.ssh
#> - name: gcr.io/gcer-public/packagetools:master
#>   args:
#>   - Rscript
#>   - -e
#>   - |-
#>     devtools::install()
#>     pkgdown::build_site()
#>   dir: repo
#> - name: gcr.io/cloud-builders/git
#>   args:
#>   - add
#>   - --all
#>   volumes:
#>   - name: ssh
#>     path: /root/.ssh
#> - name: gcr.io/cloud-builders/git
#>   args:
#>   - commit
#>   - -a
#>   - -m
#>   - "[skip travis] Build website from commit ${COMMIT_SHA}: \n$(date +\"%Y%m%dT%H:%M:%S\")"
#>   volumes:
#>   - name: ssh
#>     path: /root/.ssh
#> - name: gcr.io/cloud-builders/git
#>   args:
#>   - status
#>   volumes:
#>   - name: ssh
#>     path: /root/.ssh
#> - name: gcr.io/cloud-builders/git
#>   args:
#>   - push
#>   volumes:
#>   - name: ssh
#>     path: /root/.ssh

file.exists("cloudbuild-pkgdown.yml")

#> [1] TRUE

unlink("cloudbuild-pkgdown.yml")