Schedule a Build object via HTTP or PubSub

Schedule a Build object via HTTP or PubSub

cr_schedule_build(
  build,
  schedule,
  schedule_type = c("http", "pubsub"),
  email = cr_email_get(),
  projectId = cr_project_get(),
  ...
)

Arguments

build

A Build object

schedule

A cron schedule e.g. "15 5 * * *"

schedule_type

Whether to use HTTP or PubSub styled schedules

email

The email that will authenticate the job set via cr_email_set

projectId

The GCP project to run within usually set with cr_project_set

...

Arguments passed on to cr_schedule

region

The region usually set with cr_region_set

overwrite

If TRUE and an existing job with the same name exists, will overwrite it with the new parameters

name

Name to call your scheduled job

httpTarget

A HTTP target object HttpTarget

pubsubTarget

A Pub/Sub target object PubsubTarget such as created via cr_schedule_pubsub

description

Optionally caller-specified in CreateJob or

timeZone

Specifies the time zone to be used in interpreting schedule. If set to NULL will be "UTC". Note that some time zones include a provision for daylight savings time.

Value

cr_schedule_build returns a cloud scheduler Job object

Details

See also cr_schedule which you can use by to customise your schedule.