Give a service account the right permissions for googleCloudRunner operations

Give a service account the right permissions for googleCloudRunner operations

cr_setup_service(
  account_email,
  roles = cr_setup_role_lookup("local"),
  json = Sys.getenv("GAR_CLIENT_JSON"),
  email = Sys.getenv("GARGLE_EMAIL")
)

cr_setup_role_lookup(
  type = c("local", "cloudrun", "bigquery", "secrets", "cloudbuild", "cloudstorage",
    "schedule_agent", "run_agent")
)

Arguments

account_email
The service account email e.g. accountId@projectid.iam.gserviceaccount.com or 12345678@cloudbuild.gserviceaccount.com

roles
the roles to grant access - default is all googleCloudRunner functions

json
the project clientId JSON

email
the email of an Owner/Editor for the project

type
the role

See also