BuildTrigger Object

Configuration for an automated build in response to source repositorychanges.

BuildTrigger(
  filename = NULL,
  name = NULL,
  tags = NULL,
  build = NULL,
  ignoredFiles = NULL,
  github = NULL,
  substitutions = NULL,
  includedFiles = NULL,
  disabled = NULL,
  triggerTemplate = NULL,
  description = NULL
)

Arguments

filename
Path, from the source root, to a file whose contents is used for the build

name
User assigned name of the trigger

tags
Tags for annotation of a `BuildTrigger`

build
Contents of the build template

ignoredFiles
ignored_files and included_files are file glob matches extended with support for "**".

github
a GitHubEventsConfig object - mutually exclusive with triggerTemplate

substitutions
A named list of Build macro variables

includedFiles
If any of the files altered in the commit pass the ignored_files

disabled
If true, the trigger will never result in a build

triggerTemplate
a RepoSource object - mutually exclusive with github

description
Human-readable description of this trigger

Value

BuildTrigger object

See also