Job Schedule Object

Job Schedule Object

Job(
  name = NULL,
  description = NULL,
  schedule = NULL,
  timeZone = NULL,
  userUpdateTime = NULL,
  state = NULL,
  status = NULL,
  scheduleTime = NULL,
  lastAttemptTime = NULL,
  retryConfig = NULL,
  attemptDeadline = NULL,
  pubsubTarget = NULL,
  appEngineHttpTarget = NULL,
  httpTarget = NULL
)

Arguments

name

Name to call your scheduled job

description

Optionally caller-specified in CreateJob or

schedule

A cron schedule e.g. "15 5 * * *"

timeZone

Specifies the time zone to be used in interpreting schedule. If set to NULL will be "UTC". Note that some time zones include a provision for daylight savings time.

userUpdateTime

Output only

state

Output only

status

Output only

scheduleTime

Output only

lastAttemptTime

Output only

retryConfig

Settings that determine the retry behavior

attemptDeadline

The deadline for job attempts

pubsubTarget

A Pub/Sub target object PubsubTarget such as created via cr_schedule_pubsub

appEngineHttpTarget

App Engine HTTP target

httpTarget

A HTTP target object HttpTarget

Value

Job object

Details

Configuration for a job.The maximum allowed size for a job is 100KB.

See also