Make a Cloud Build object out of a cloudbuild.yml file

This creates a Build object via the standard cloudbuild.yaml format

cr_build_make(
  yaml,
  source = NULL,
  timeout = NULL,
  images = NULL,
  artifacts = NULL,
  options = NULL,
  substitutions = NULL,
  availableSecrets = NULL,
  logsBucket = NULL
)

Arguments

yaml
A Yaml object created from cr_build_yaml or a file location of a .yaml/.yml cloud build file

source
A Source object specifying the location of the source files to build, usually created by cr_build_source

timeout
Amount of time that this build should be allowed to run, to second

images
A list of images to be pushed upon the successful completion of all build

artifacts
Artifacts that may be built via cr_build_yaml_artifact

options
Options to pass to a Cloud Build

substitutions
Substitutions data for `Build` resource

availableSecrets
Secret Manager objects built by cr_build_yaml_secrets

logsBucket
The gs:// location of a bucket to put logs in

See also

Examples

cloudbuild <- system.file("cloudbuild/cloudbuild.yaml",
                           package = "googleCloudRunner")
cr_build_make(cloudbuild)

#> ==CloudBuildObject==
#> steps:
#> - name: gcr.io/cloud-builders/docker
#>   id: Docker Version
#>   args: version
#> - name: alpine
#>   id: Hello Cloud Build
#>   args:
#>   - echo
#>   - Hello Cloud Build
#> - name: rocker/r-base
#>   id: Hello R
#>   args:
#>   - Rscript
#>   - -e
#>   - paste0('1 + 1 = ', 1+1)